% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R, R/MakeCorrelationNetwork.R
\name{BindMatrices}
\alias{BindMatrices}
\title{Bind Matrices}
\usage{
BindMatrices(cluster_list, correlation_matrix)

BindMatrices(cluster_list, correlation_matrix)
}
\arguments{
\item{cluster_list}{A list of clusters to generate adjacency matrices.}

\item{correlation_matrix}{A correlation matrix to align with the adjacency matrix.}
}
\value{
The CCCN matrix, the intersect between the correlation matrix and the ordered adjacency matrix.

The CCCN matrix, the intersect between the correlation matrix and the ordered adjacency matrix.
}
\description{
This function orders the adjacency matrix, finds the intersect with the correlation matrix (cccn_matrix), and puts NAs on the diagonal.

This function orders the adjacency matrix, finds the intersect with the correlation matrix (cccn_matrix), and puts NAs on the diagonal.
}
\examples{
BindMatrices(cluster_list, correlation_matrix)
BindMatrices(cluster_list, correlation_matrix)
}
