% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayCrosstalkNetwork.R
\name{bioplanet.innit}
\alias{bioplanet.innit}
\title{Helper function to take information from a bioplanet file and format it in the desired way. Assumes that the function was downloaded from excel.}
\usage{
bioplanet.innit(fn)
}
\arguments{
\item{filename}{The name of the bioplanet pathway .csv file}
}
\value{
A formatted version of the bioplanet csv file
}
\description{
Helper function to take information from a bioplanet file and format it in the desired way. Assumes that the function was downloaded from excel.
}
\examples{
print("TO DO")
}
