% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{brca_CCCN_data}
\alias{brca_CCCN_data}
\title{Output of MakeCorrelationNetwork on the BRCA data}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(brca_CCCN_data)
}
\description{
Output of MakeCorrelationNetwork on the BRCA data
}
\keyword{internal}
