% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{brca_clusterlist_data}
\alias{brca_clusterlist_data}
\title{BRCA Cluster List Data}
\format{
An object of class \code{list} of length 3.
}
\usage{
brca_clusterlist_data
}
\description{
Output of MakeClusterList on the BRCA data
}
\keyword{internal}
