% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{BRCA_genemania.edges}
\alias{BRCA_genemania.edges}
\title{BRCA Genemania Edges}
\format{
An object of class \code{data.frame} with 18836 rows and 4 columns.
}
\usage{
BRCA_genemania.edges
}
\description{
Edges from the Genemania database for the genes in the BRCA dataset
}
\keyword{internal}
