% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{BRCA_PCN.data}
\alias{BRCA_PCN.data}
\title{BRCA PCN Data}
\format{
An object of class \code{list} of length 3.
}
\usage{
BRCA_PCN.data
}
\description{
Output of BuildPathwayCrosstalkNetwork on the BRCA data
}
\keyword{internal}
