% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{BRCA_stringdb.edges}
\alias{BRCA_stringdb.edges}
\title{BRCA STRINGdb Edges}
\format{
An object of class \code{data.frame} with 51286 rows and 4 columns.
}
\usage{
BRCA_stringdb.edges
}
\description{
Edges from the STRINGdb database for the genes in the BRCA dataset
}
\keyword{internal}
