% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPPINetwork.R
\name{BuildClusterFilteredNetwork}
\alias{BuildClusterFilteredNetwork}
\title{Build PPI Network}
\usage{
BuildClusterFilteredNetwork(
  stringdb.edges,
  genemania.edges,
  gene.cccn,
  db.filepaths = c(),
  ppi.network.name = "combined.PPIs",
  cfn.name = "cfn"
)
}
\arguments{
\item{stringdb.edges}{Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database}

\item{genemania.edges}{GeneMANIA network of protein-protein interactions from the genes of study; defaults to NA}

\item{gene.cccn}{A matrix showing strength of relationships between proteins using the common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))}

\item{db.filepaths}{A vector of paths to the additional ppi network files; defaults to an empty vector}

\item{ppi.network.name}{Desired name for the output protein-protein interaction network using all entered database input; defaults to ppi.network}

\item{cfn.name}{Desired name of the output cluster filtered network}

\item{ppi.network}{A dataframe representing how strongly proteins are known to interact based on previous research contained within provided database networks}
}
\value{
A dataframe representing how strongly proteins are known to interact based on previous research contained within provided database networks

A version of ppi.network with only the edges that exist in gene.cccn and have non-zero weights
}
\description{
Uses all inputted database information on protein-protein interactions in order to create a combined PPI-network.
}
\examples{
ex.string <- ex.stringdb.edges
ex.gm <- ex.genemania.edges
BuildPPINetwork(stringdb.edges = ex.string, genemania.edges = ex.gm, ppi.network.name = "ex.ppi.network")
utils::head(ex.ppi.network)

#' Filter PPI Network by Known Clusters

All edges in the PPI network are checked against the cocluster correlation network to ensure existence and a non-zero weight.
If either of these conditions are not met, then it will be removed from the list of PPI edges.
This new, cluster filtered network is then assigned to the global namespace.

ClusterFilteredNetwork(ex.gene.cccn, ex.ppi.network, cfn.name = "ex.cfn")
utils::head(ex.cfn)
}
