% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPPINetwork.R
\name{BuildPPINetwork}
\alias{BuildPPINetwork}
\title{Build PPI Network}
\usage{
BuildPPINetwork(removeme)
}
\arguments{
\item{stringdb.edges}{Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database}

\item{genemania.edges}{GeneMANIA network of protein-protein interactions from the genes of study; defaults to NA}

\item{db.filepaths}{A vector of paths to the additional ppi network files; defaults to an empty vector}

\item{ppi.network.name}{Desired name for the output protein-protein interaction network using all entered database input; defaults to ppi.network}
}
\value{
A dataframe representing how strongly proteins are known to interact based on previous research contained within provided database networks
}
\description{
Uses all inputted database information on protein-protein interactions in order to create a combined PPI-network.
}
\examples{
ex.string <- ex.stringdb.edges
ex.gm <- ex.genemania.edges
BuildPPINetwork(stringdb.edges = ex.string, genemania.edges = ex.gm, ppi.network.name = "ex.ppi.network")
utils::head(ex.ppi.network)
}
