% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPPIEdges.R
\name{cccn_to_nodenames}
\alias{cccn_to_nodenames}
\title{Pulls nodenames from the cccn_matrix}
\usage{
cccn_to_nodenames(cccn_matrix)
}
\arguments{
\item{cccn_matrix}{dataframe of dataframes that represent the common clusters from the three distance calculations' clusters}
}
\value{
list of the names of the genes
}
\description{
This helper function pulls the gene names from the cccn_matrix into a list 'nodenames'
}
\examples{
cccn_to_nodenames(cccn_matrix)
}
