% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{clean.string}
\alias{clean.string}
\title{Clean String}
\usage{
clean.string(string)
}
\arguments{
\item{string}{A string with special characters}
}
\value{
A string with no special character
}
\description{
A helper function to remove all special characters from a string
}
