% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterFilteredNetwork.R
\name{ClusterFilteredNetwork}
\alias{ClusterFilteredNetwork}
\title{ClusterFilteredNetwork}
\usage{
ClusterFilteredNetwork()
}
\value{
A data frame of correlation edges.
}
\description{
This function processes correlation edges from a given correlation matrix
}
\examples{
cccn.cfn.tools:::ClusterFilteredNetwork()
}
