% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayCrosstalkNetwork.R
\name{ClusterPathwayEvidence}
\alias{ClusterPathwayEvidence}
\title{Cluster Pathway Evidence}
\usage{
ClusterPathwayEvidence(cluster, pathway, p.list)
}
\arguments{
\item{cluster}{A cluster of genes/proteins (cannot be PTMs)}

\item{pathway}{A pathway of genes/proteins}

\item{p.list}{A list of pathways to calculate the # of pathways a gene, k, is in}
}
\value{
A float value representing cluster pathway evidence between a cluster and pathway
}
\description{
Helper function for obtaining cluster pathway evidence given a Cluster and Pathway
}
