% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{CorrelationNetwork}
\alias{CorrelationNetwork}
\title{Generate Correlation Network}
\usage{
CorrelationNetwork(cccn_matrix)
}
\arguments{
\item{cccn_matrix}{The CCCN matrix, the intersect between the correlation matrix and the ordered adjacency matrix produced from \code{BindMatrices}.}
}
\value{
An igraph object representing the correlation network.
}
\description{
This function takes the cccn_matrix, replaces the NA values with zeroes, and turns it into an igraph object.
}
\examples{
CorrelationNetwork(cccn_matrix)
}
