% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean.R
\name{dates.to.ptms}
\alias{dates.to.ptms}
\title{Fix Excel}
\usage{
dates.to.ptms(vector)
}
\arguments{
\item{vector}{A character vector (ideally a column of a table)}
}
\value{
A fixed version of the character vector
}
\description{
A helper function that tries to identify dates and turn them into the names of common ptms such as 1-Sep -> SEPT1
NOTE: This should probably be called over an entire table for user simplicity
}
