% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.bioplanet}
\alias{ex.bioplanet}
\title{Bioplanet Example}
\format{
Dataframe with 100 rows and 4 columns:
\describe{
\item{}{}
...
}
}
\source{
source info (or \code{"manually created"})
}
\usage{
ex.bioplanet
}
\description{
(description)
}
\examples{
data(ex.bioplanet)
head(ex.bioplanet)
}
\keyword{datasets}
