% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.cccn_matrix}
\alias{ex.cccn_matrix}
\title{CoCluster Correlation Network}
\format{
Square matrix with 19 rows and columns:
\describe{
\item{rows/columns}{Gene}
\item{values}{Strength of correlation}
}
}
\source{
"created by MakeCorrelationNetwork"
}
\usage{
ex.cccn_matrix
}
\description{
Network of relationships between genes based on clustering of ptms from the three distance metrics.
}
\examples{
data(ex.cccn_matrix)
head(ex.cccn_matrix)
}
\keyword{datasets}
