% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Examples.R
\name{ex.ClusterFilteredNetwork}
\alias{ex.ClusterFilteredNetwork}
\title{Cluster Filtered Network Example}
\usage{
ex.ClusterFilteredNetwork(pBound = 0.5, nBound = -0.5, mode = "lower")
}
\description{
Cluster Filtered Network Example
}
\keyword{internal}
