% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.clusters.common}
\alias{ex.clusters.common}
\title{List of Common Clusters Example}
\format{
List of 40 lists:
\describe{
\item{list}{Common cluster between the three distance metrics}
}
}
\source{
source info (or \code{"manually created"})
}
\usage{
ex.clusters.common
}
\description{
(description)
}
\examples{
data(ex.common.clusters)
head(ex.common.clusters)
}
\keyword{internal}
