% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.common.clusters}
\alias{ex.common.clusters}
\title{List of Common Clusters Example}
\format{
List of 40 lists:
\describe{
\item{list}{Common cluster between the three distance metrics}
}
}
\source{
"Created in MakeCorrelationNetwork, FindCommonClusters using 3 T-SNE lists"
}
\usage{
ex.common.clusters
}
\description{
Clusters that contain PTMs that are found to cocluster in all 3 T-SNE lists.
}
\examples{
head(ex.common.clusters)
}
\keyword{internal}
