% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.CPE.matrix}
\alias{ex.CPE.matrix}
\title{CPE matrix example}
\format{
Matrix of 40 rows and 9 columns
\describe{
\item{rows}{Clusters}
\item{columns}{Pathways}
\item{values}{Cluster Pathway Evidence}
}
}
\source{
"Made in PathwayCrosstalkNetwork"
}
\usage{
ex.CPE.matrix
}
\description{
Scores that represents how likely it is that a pathway interacts with a given cluster. Matrix of Clusters x Pathways.
}
\examples{
ex.CPE.matrix[c(1,2) , c(1,2)]
}
\keyword{internal}
