% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Examples.R
\name{ex.find_ppi_edges}
\alias{ex.find_ppi_edges}
\title{Find PPI Edges Example}
\usage{
ex.find_ppi_edges(cccn_matrix, gmfilepath = "genemania-interactions.txt")
}
\description{
Find PPI Edges Example
}
\keyword{internal}
