% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.funckey}
\alias{ex.funckey}
\title{Funckey}
\format{
Dataframe of 24339 rows and 8 columns:
\describe{
\item{Gene.Name}{Name of the gene}
\item{Approved.Name}{Name of the protein}
\item{Hugo.Gene.Family}{Family of the gene}
\item{HPRD.Function}{Function of the protein}
\item{nodeType}{Type of protein}
\item{Domains}{Known domains of protein}
\item{Compartment}{Location of the protein}
\item{Compartment.Overview}{Alternate location of the protein}
}
}
\source{
""
}
\usage{
ex.funckey
}
\description{
DESCRIPTION
}
\examples{
head(ex.funckey)
}
\keyword{internal}
