% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.gene.cccn}
\alias{ex.gene.cccn}
\title{Gene CoCluster Network example}
\format{
Matrix of 74 rows and columns
\describe{
\item{rows}{Genes in PTM cccn}
\item{columns}{Genes in PTM cccn}
\item{values}{Sum of submatrices whose PTMs are on said gene}
}
}
\source{
"Made in MakeCorrelationNetwork"
}
\usage{
ex.gene.cccn
}
\description{
A version of the PTM cccn that works at the Gene level instead of the PTm level
}
\examples{
ex.gene.cccn[c(1,2) , c(1,2)]
}
\keyword{internal}
