% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.gm.edges}
\alias{ex.gm.edges}
\title{Full GeneMANIA Edges Example}
\format{
Dataframe of 601,552 rows and 10 columns:
\describe{
\item{data.type}{Type of interaction; typically co-expression}
\item{highlight}{UNKNOWN}
\item{interaction}{UNKNOWN}
\item{name}{The names of the two genes interacting as well as the data type}
\item{networks}{Source of information}
\item{normalized.max.weight}{UNKNOWN}
\item{raw.weights}{A number weighting the strength of the interaction}
\item{selected}{UNKNOWN}
\item{shared.interaction}{UNKNOWN}
\item{shared.name}{UNKNOWN}
...
}
}
\source{
\url{https://apps.cytoscape.org/apps/genemania}
}
\usage{
ex.gm.edges
}
\description{
Network of interactions from the GeneMANIA database.
}
\examples{
data(ex.gm.edges)
head(ex.gm.edges)
}
\keyword{datasets}
