% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.gm.network}
\alias{ex.gm.network}
\title{Full GeneMANIA Network Example}
\format{
Dataframe of 601,552 rows and 3 columns:
\describe{
\item{}{}
...
}
}
\source{
\url{https://apps.cytoscape.org/apps/genemania}
}
\usage{
ex.gm.network
}
\description{
Network of relationships between genes based on information from the GeneMANIA database.
}
\examples{
data(ex.gm.network)
head(ex.gm.network)
}
\keyword{datasets}
