% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.gm.nodes}
\alias{ex.gm.nodes}
\title{GeneMANIA Nodes Example}
\format{
Dataframe of 2762 rows and 16 columns:
\describe{
\item{}{}
...
}
}
\source{
\url{https://apps.cytoscape.org/apps/genemania}
}
\usage{
ex.gm.nodes
}
\description{
Dataframe from GeneMANIA relating GeneMANIA IDs to common gene names.
}
\examples{
data(ex.gm.nodes)
head(ex.gm.nodes)
}
\keyword{datasets}
