% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.gm_edges}
\alias{ex.gm_edges}
\title{GeneMANIA Empty Edges Example}
\format{
Dataframe of 0 rows and 4 columns:
\describe{
\item{}{}
...
}
}
\source{
\url{https://genemania.org}
}
\usage{
ex.gm_edges
}
\description{
Empty GeneMANIA structure. TODO: fix.
}
\examples{
data(ex.gm_edges)
head(ex.gm_edges)
}
\keyword{datasets}
