% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.Jaccard.Full}
\alias{ex.Jaccard.Full}
\title{Jaccard Example}
\format{
Matrix with 36 rows and 3 columns:
\describe{
\item{V1}{Pathway One}
\item{V2}{Pathway Two}
\item{jaccard.values}{Jaccard similarity between the two pathways}
}
}
\source{
"created by PathwayCrosstalkNetwork"
}
\usage{
ex.Jaccard.Full
}
\description{
A matrix measuring the Jaccard similarity between pairs ofpathways
}
\examples{
head(ex.Jaccard.Full)
}
\keyword{internal}
