% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.list.common}
\alias{ex.list.common}
\title{List of Common Clusters Example}
\format{
List of 12 variables:
\describe{
\item{}{}
...
}
}
\source{
source info (or \code{"manually created"})
}
\usage{
ex.list.common
}
\description{
(description)
}
\examples{
data(ex.list.common)
head(ex.list.common)
}
\keyword{datasets}
