% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.nodenames}
\alias{ex.nodenames}
\title{Nodenames Example}
\format{
Dataframe of 10 rows and 1 column:
\describe{
\item{Gene.Names}{Name of the gene}
}
}
\source{
"created by MakeDBInput"
}
\usage{
ex.nodenames
}
\description{
List of gene names in dataset.
}
\examples{
data(ex.nodenames)
head(ex.nodenames)
}
\keyword{internal}
