% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.pathways.list}
\alias{ex.pathways.list}
\title{Pathways List Example}
\format{
List of 1657 elements:
\describe{
\item{pathway}{A pathway name with the list of genes involved}
}
}
\source{
source info (or \code{"manually created"})
}
\usage{
ex.pathways.list
}
\description{
(description)
}
\examples{
data(ex.pathways.list)
head(ex.pathways.list)
}
\keyword{internal}
