% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Examples.R
\name{ex.ProcessGMEdgefile}
\alias{ex.ProcessGMEdgefile}
\title{Process GeneMANIA example}
\usage{
ex.ProcessGMEdgefile(epath, npath, ex.nodenames, ex.gm.network)
}
\description{
Process GeneMANIA example
}
\keyword{internal}
