% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.ptm.cccn}
\alias{ex.ptm.cccn}
\title{Filtered Correlation Matrix example}
\format{
Matrix of 329 rows and columns
\describe{
\item{rows}{Select PTMs}
\item{columns}{Select PTMs}
\item{values}{Same as in Correlation Matrix}
}
}
\source{
"Made in MakeCorrelationNetwork"
}
\usage{
ex.ptm.cccn
}
\description{
A version of the Correlation Matrix filtered by coclustering PTMs
}
\examples{
ex.ptm.cccn[c(1,2) , c(1,2)]
}
\keyword{internal}
