% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.ptm.correlation.matrix}
\alias{ex.ptm.correlation.matrix}
\title{Correlation Matrix Example}
\format{
Square matrix of 100 rows and columns:
\describe{
\item{}{}
...
}
}
\source{
source info (or \code{"manually created"})
}
\usage{
ex.ptm.correlation.matrix
}
\description{
Matrix relating the correlation values between ptms based on Spearman Similarity.
}
\examples{
data(ex.ptm.correlation.matrix)
head(ex.ptm.correlation.matrix)
}
\keyword{datasets}
