% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.PTPedgelist}
\alias{ex.PTPedgelist}
\title{PTP Edgelist Example}
\format{
Dataframe of 1874 rows and 4 columns:
\describe{
\item{Source}{}
\item{target}{}
\item{Jaccard_weight}{}
\item{PTP_inter_evidence}{}
}
}
\source{
""
}
\usage{
ex.PTPedgelist
}
\description{
DESCRIPTION
}
\examples{
data(ex.PTPedgelist)
head(ex.PTPedgelist)
}
\keyword{internal}
