% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.tsne.matrices}
\alias{ex.tsne.matrices}
\title{TSNE matrices Example}
\format{
List of 3 lists of 6 (Euclidean and SED) or 88 (Spearman) dataframes of varying sizes:
\describe{
\item{}{}
...
}
}
\source{
source info (or \code{"manually created"})
}
\usage{
ex.tsne.matrices
}
\description{
List of of lists of dataframes of ptm positions in 3D space to show their relationships with each other based on distance.
}
\examples{
data(ex.tsne.matrices)
head(ex.tsne.matrices)
}
\keyword{datasets}
