% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_adj_consensus}
\alias{ex_adj_consensus}
\title{Adjacency Consensus Matrix}
\format{
Dataframe of 908 rows and columns:
\describe{
\item{rows}{PTMs}
\item{columns}{PTMs}
\item{values}{1 if PTMs are in the same cluster (pls confirm)}
}
}
\source{
"Produced by MakeClusterList"
}
\usage{
ex_adj_consensus
}
\description{
A matrix containing values of 0s and 1s depending on if the PTMs are in a cocluster with other PTMs.
}
\examples{
head(ex_adj_consensus)
}
\keyword{internal}
