% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_bioplanet}
\alias{ex_bioplanet}
\title{Bioplanet}
\format{
Dataframe of 196 rows and 4 columns
\describe{
\item{PATHWAY_ID}{The pathway a gene belongs to}
\item{PATHWAY_NAME}{The name of the pathway a gene belongs to}
\item{GENE_ID}{A unique gene ID}
\item{GENE_SYMBOL}{An abrivation for the name of the gene}
}
}
\source{
"https://www.frontiersin.org/journals/pharmacology/articles/10.3389/fphar.2019.00445/full"
}
\usage{
ex_bioplanet
}
\description{
A database file from the bioplanet website
}
\examples{
head("ex_bioplanet")
}
\keyword{internal}
