% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_cfn}
\alias{ex_cfn}
\title{Cfn}
\format{
Dataframe of 196 rows and 4 columns
\describe{
\item{source}{First gene interacting}
\item{target}{Second gene interacting}
\item{interaction}{The type of interaction}
\item{Weight}{Weight from Combined PPI}
}
}
\source{
"Produced by BuildClusterFilteredNetwork"
}
\usage{
ex_cfn
}
\description{
The Cluster Filtered Network is a filtered version of the Combined PPIs but only for proteins found to interact via t-SNE clusters
}
\examples{
head("ex_cfn")
}
\keyword{internal}
