% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_combined_ppi}
\alias{ex_combined_ppi}
\title{Combined PPIs}
\format{
Dataframe of 8849 rows and 4 columns
\describe{
\item{source}{First gene interacting}
\item{target}{Second gene interacting}
\item{interaction}{The type of interaction}
\item{Weight}{The normalized weight of the interaction between two genes from Genemania and StringDB}
}
}
\source{
"Produced by BuildClusterFilteredNetwork"
}
\usage{
ex_combined_ppi
}
\description{
Combines all Protein to Protein interaction data and combines their weights into one score from a range of 0 to 1
}
\examples{
head("ex_combined_ppi")
}
\keyword{internal}
