% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_common_clusters}
\alias{ex_common_clusters}
\title{Common Clusters}
\format{
List of 140 lists:
\describe{
\item{Cluster id}{Common clusters between the three distance metrics}
}
}
\source{
"Produced by MakeClusterList"
}
\usage{
ex_common_clusters
}
\description{
Clusters that contain PTMs that are found to cocluster in all 3 T-SNE lists.
}
\examples{
ex_common_clusters[0:3]
}
\keyword{internal}
