% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_full_ptm_table}
\alias{ex_full_ptm_table}
\title{Full PTM Table Example}
\format{
Dataframe of 9215 rows and 70 columns:
\describe{
\item{row}{Number of PTMs}
\item{columns}{Names of drugs that were used on PTMs}
\item{values}{Expression as found by mass spectrometry}
}
}
\source{
"Mass Spectrometry Data"
}
\usage{
ex_full_ptm_table
}
\description{
Post-translational modification data under certain environmental conditions.
}
\examples{
head(ex_full_ptm_table)
}
\keyword{internal}
