% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_gene_cccn_edges}
\alias{ex_gene_cccn_edges}
\title{Gene CCCN Edgelist}
\format{
Data Frame of 2279 rows and 4 columns
\describe{
\item{source}{A Gene}
\item{target}{A Gene}
\item{Weight}{Value from -1 to 1}
\item{interaction}{Either correlation, positive correlation, or negative correlation}
}
}
\source{
"Produced by MakeCorrelationNetwork"
}
\usage{
ex_gene_cccn_edges
}
\description{
Gene Co-Cluster Correlation Network as an Edgelist
}
\examples{
head(ex_gene_cccn_edges)
}
\keyword{internal}
