% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_gene_cccn_nodes}
\alias{ex_gene_cccn_nodes}
\title{Gene list (nodes)}
\format{
Character Vector of length 390
\describe{
\item{String}{Gene}
}
}
\source{
"Produced by MakeCorrelationNetwork"
}
\usage{
ex_gene_cccn_nodes
}
\description{
List of Genes gotten from PTMs
}
\examples{
ex_gene_cccn_nodes[0:5]
}
\keyword{internal}
