% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_genemania_edges}
\alias{ex_genemania_edges}
\title{Genemania Edges}
\format{
Dataframe of 5 rows and 4 columns
\describe{
\item{source}{First gene interacting}
\item{target}{Second gene interacting}
\item{interaction}{The type of interaction}
\item{Weight}{The weight of the interaction between these two genes gotten from the Genemania database}
}
}
\source{
"Produced by GetGeneMANIA.edges"
}
\usage{
ex_genemania_edges
}
\description{
Edges from gene_cccn_nodes that are also found within the Genemania database
}
\examples{
head(ex_genemania_edges)
}
\keyword{internal}
