% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_pathway_crosstalk_network}
\alias{ex_pathway_crosstalk_network}
\title{Pathway Crosstalk Network}
\format{
Dataframe of 37 rows and 4 columns
\describe{
\item{source}{The name of a pathway interacting}
\item{target}{The name of the pathway the source interacts with}
\item{Weight}{Weight from Combined PPI}
\item{interaction}{The type of interaction}
}
}
\source{
"Produced by BuildPathwayCrosstalkNetwork"
}
\usage{
ex_pathway_crosstalk_network
}
\description{
Specific genes from the bioplanet file that we found in Common Clusters and their respective cell signaling pathway relationships
}
\examples{
head("ex_pathway_crosstalk_network")
}
\keyword{internal}
