% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_pathways_list}
\alias{ex_pathways_list}
\title{Pathways list}
\format{
list of 9 character vectors
\describe{
\item{Pathway name}{Genes in the pathway}
}
}
\source{
"Produced by BuildPathwayCrosstalkNetwork"
}
\usage{
ex_pathways_list
}
\description{
A list of pathway names and the genes that make them up
}
\examples{
head("ex_pathways_list")
}
\keyword{datasets}
