% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_PCNedgelist}
\alias{ex_PCNedgelist}
\title{PCN Edge List}
\format{
Dataframe of 19 rows and 4 columns
\describe{
\item{source}{The name of a pathway interacting}
\item{target}{The name of the pathway the source interacts with}
\item{pathway_Jaccard_similarity}{A weight}
\item{PTM_cluster_evidence}{A weight}
}
}
\source{
"Produced by BuildPathwayCrosstalkNetwork"
}
\usage{
ex_PCNedgelist
}
\description{
The Pathway Crosstalk Network but with direct values for both weight types
}
\examples{
head("ex_PCNedgelist")
}
\keyword{internal}
