% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_ptm_cccn_edges}
\alias{ex_ptm_cccn_edges}
\title{PTM CCCN Edgelist}
\format{
Data Frame of 5710 rows and 4 columns
\describe{
\item{source}{A PTM}
\item{target}{A PTM}
\item{Weight}{Value from -1 to 1}
\item{interaction}{Either correlation, positive correlation, or negative correlation}
}
}
\source{
"Produced by MakeCorrelationNetwork"
}
\usage{
ex_ptm_cccn_edges
}
\description{
PTM Co-Cluster Correlation Network as an Edgelist
}
\examples{
head(ex_ptm_cccn_edges)
}
\keyword{internal}
