% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_ptm_correlation_matrix}
\alias{ex_ptm_correlation_matrix}
\title{Correlation Matrix}
\format{
Square matrix of 908 rows and columns:
\describe{
\item{rows and columns}{Ptm names}
\item{values}{Strength of correlation}
}
}
\source{
"Produced by MakeClusterList"
}
\usage{
ex_ptm_correlation_matrix
}
\description{
Matrix relating the correlation values between ptms based on Spearman Similarity.
}
\examples{
head(ex_ptm_correlation_matrix)
}
\keyword{internal}
