% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_small_ptm_table}
\alias{ex_small_ptm_table}
\title{Small PTM Table Example}
\format{
Dataframe of 497 rows and 72 columns:
\describe{
\item{row}{Number of PTMs}
\item{columns}{Names of drugs that were used on PTMs}
\item{values}{Expression as found by mass spectrometry}
}
}
\source{
"An .rda version of phospho_cleaned_mapped.txt found in inst/extdata"
}
\usage{
ex_small_ptm_table
}
\description{
Post-translational modification data under certain environmental conditions.
}
\examples{
head(ex_small_ptm_table)
}
\keyword{internal}
