% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_stringdb_edges}
\alias{ex_stringdb_edges}
\title{STRINGdb Edges}
\format{
Dataframe of 8844 rows and 4 columns:
\describe{
\item{source}{First gene interacting}
\item{target}{Second gene interacting}
\item{interaction}{The type of interaction}
\item{Weight}{The weight of the interaction between these two genes as described by STRINGdb}
}
}
\source{
"Produced by GetSTRINGdb.edges"
}
\usage{
ex_stringdb_edges
}
\description{
List of edges between genes in dataset.
}
\examples{
head(ex_stringdb_edges)
}
\keyword{internal}
