% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_tiny_ptm_table}
\alias{ex_tiny_ptm_table}
\title{Tiny PTM Table Example}
\format{
Dataframe of 125 rows and 7 columns:
\describe{
\item{row}{Number of PTMs}
\item{columns}{Names of drugs that were used on PTMs}
\item{values}{Expression as found by mass spectrometry}
}
}
\source{
"Sub-data frame of the small PTM table"
}
\usage{
ex_tiny_ptm_table
}
\description{
Post-translational modification data under certain environmental conditions.
}
\examples{
head(ex_small_ptm_table)
}
\keyword{internal}
