% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPPIEdges.R
\name{find_ppi_edges}
\alias{find_ppi_edges}
\title{Find PPI Edges}
\usage{
find_ppi_edges(cccn_matrix, gmfilename = "genemania-interactions.txt")
}
\arguments{
\item{cccn_matrix}{dataframe of dataframes that represent the common clusters from the three distance calculations' clusters}

\item{gmfilename}{The path to the GeneMania file initialized to "genemania-interactions.txt".}
}
\value{
A data frame of combined edges from STRINGdb and GeneMANIA.
}
\description{
This function finds protein-protein interaction edges by combining STRINGdb and GeneMANIA databases.
}
\examples{
find_ppi_edges(cccn_matrix, gmfilename = 'genemania/genemania-interactions.txt')
}
