% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherPPIData.R
\name{formatKinsubTable}
\alias{formatKinsubTable}
\title{Format Kinsub Table}
\usage{
formatKinsubTable(kinasesubstrate.filename = "Kinase_Substrate_Dataset.txt")
}
\arguments{
\item{kinasesubstrate.filename}{The path to the kinase substrate database file from https://www.phosphosite.org/staticDownloads}
}
\value{
An edgelist filtered by the gene cccn and nodenames
}
\description{
Include kinase substrate dataset from PhosphoSitePlus https://www.phosphosite.org/staticDownloads
}
