% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{function_key}
\alias{function_key}
\title{Function Key Example}
\format{
Data frame with 24339 rows and 8 columns:
\describe{
\item{Gene.Name}{HGNC gene symbol.}
\item{Approved.Name}{Approved gene name.}
\item{Hugo.Gene.Family}{Gene family annotation.}
\item{HPRD.Function}{Functional annotation from HPRD.}
\item{nodeType}{Node classification used for Cytoscape styling.}
\item{Domains}{Protein domain annotations.}
\item{Compartment}{Subcellular compartment labels.}
\item{Compartment.Overview}{Simplified compartment grouping.}
}
}
\source{
\code{FunctionKey.txt} processed into a package data object.
}
\usage{
data(function_key)
}
\description{
Gene-level annotation data used to build Cytoscape node metadata.
}
\examples{
head(function_key)
}
\keyword{datasets}
