% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{GenerateAndConstructptmsNetwork}
\alias{GenerateAndConstructptmsNetwork}
\title{Generate and Construct All PTMs Network}
\usage{
GenerateAndConstructptmsNetwork(ptmtable, keeplength = 2, output_dir = "plots")
}
\arguments{
\item{ptmtable}{A data frame containing PTM data.}

\item{keeplength}{An integer specifying the minimum length of common elements to keep. Default is 2.}

\item{output_dir}{A string specifying the output directory for saving plots. Default is "plots".}
}
\value{
\code{eu.sp.sed.ptms}, the list of clusters.
}
\description{
This function generates and constructs the PTMs network from the ptmtable and previously calculated variables.
}
\examples{
GenerateAndConstructptmsNetwork(ptmtable)
}
