% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{get.GM.edgefile}
\alias{get.GM.edgefile}
\title{Loads and filters the GeneMania file given a vector of gene names.}
\usage{
get.GM.edgefile(gmfilename, nodenames)
}
\arguments{
\item{gmfilename}{The path to the GeneMania file.}

\item{nodenames}{A vector containing the names of the relevant genes.}
}
\value{
A data frame with the relevant GeneMania data.
}
\description{
This helper function loads the GeneMania file
and filters out required nodes.
}
\examples{
get.GM.edgefile(gmfilename, nodenames)
}
