% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPPIEdges.R
\name{get.GM.edges}
\alias{get.GM.edges}
\title{Loads and filters the GeneMania file given a vector of gene names.}
\usage{
get.GM.edges(nodenames)
}
\arguments{
\item{nodenames}{A vector containing the names of the relevant genes.}
}
\value{
A data frame with the relevant GeneMania data.
}
\description{
This helper function loads the GeneMania file from the web
and filters out required nodes.
}
\examples{
get.GM.edges(nodenames)
}
