% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeClusterList.R
\name{GetRtsne}
\alias{GetRtsne}
\title{Helper function for Rtsne since parameters are set}
\usage{
GetRtsne(table, iter = 5000)
}
\arguments{
\item{table}{The matrix that Rtsne needs to be performed on. It's called table because a lot of operations contain the word "matrix" so it was called something else to not be confusing.}

\item{iter}{Number of iterations, defaults 5000}
}
\value{
A matrix where each row has 3 dimensionial t-SNE coords for a given matrix.
}
\description{
This function is made to help people new to the Rtsne function understand parameters and organize code. All it does is call Rtsne with set parameters. If you understand R code, feel free to modify the parameters and resource the function.
}
