% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherPPIData.R
\name{GetSTRINGdb}
\alias{GetSTRINGdb}
\title{Get STRINGdb PPI data}
\usage{
GetSTRINGdb(
  gene.cccn,
  stringdb.name = "stringdb.edges",
  nodenames.name = "nodenames"
)
}
\arguments{
\item{gene.cccn}{A matrix showing strength of relationships between proteins using common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))}

\item{stringdb.name}{Desired name for the output STRINGdb data frame; defaults to "stringdb.edges"}

\item{nodenames.name}{Desired name for list of gene names; defaults to nodenames}
}
\value{
Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database and a list of gene names
}
\description{
This function finds protein-protein interaction weights by consulting utilizing the STRINGdb database.
}
\details{
STRINGdb is required for this function. To download, run:
if (!require("BiocManager", quietly = TRUE)) install.packages("BiocManager")
BiocManager::install("STRINGdb")
The full example takes ~10 minutes to load, so it has been commented out and the results are displayed.
}
\examples{
# GetSTRINGdb(ex.gene.cccn, 'ex.stringdb.edges', 'ex.nodenames')
utils::head(ex.stringdb.edges)
utils::head(ex.nodenames)
}
