% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{\%w/o\%}
\alias{\%w/o\%}
\title{Set difference for vectors using custom infix operator}
\usage{
x \%w/o\% y
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector to compare.}
}
\value{
Elements of x that are not in y.
}
\description{
Returns the elements in vector x that are not in vector y.
}
\examples{
1:5 \%w/o\% c(2, 4)
letters[1:5] \%w/o\% c("b", "d")
}
