% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{list.common}
\alias{list.common}
\title{Finds common elements between clusters in three lists.}
\usage{
list.common(list1, list2, list3, keeplength = 2)
}
\arguments{
\item{list1}{A list of ptms.}

\item{list2}{A list of ptms.}

\item{keeplength}{Minimum size of intersections to keep.}
}
\value{
A list of common clusters.
}
\description{
Finds common elements between clusters in three lists.
}
\examples{
list.common(list1, list2, list3 keeplength = 2)
}
