% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPPIEdges.R
\name{make_db_input}
\alias{make_db_input}
\title{Make file for GeneMania input}
\usage{
make_db_input(cccn_matrix, file.path.name = "db_nodes.txt")
}
\arguments{
\item{cccn_matrix}{dataframe of dataframes that represent the common clusters from the three distance calculations' clusters}
}
\value{
A file with all of the gene names which can be copy and pasted into the GeneMania website, data frame of the names of the genes
}
\description{
This function outputs a file the user can take to create the GeneMania edgefile.
}
\examples{
cccn.cfn.tools:::ex.make_db_input(ex.cccn_matrix)
}
