% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeClusterList.R
\name{MakeClusterList}
\alias{MakeClusterList}
\title{Build Clusters Based on PTM Data Under Like Conditions}
\usage{
MakeClusterList(
  ptmtable,
  keeplength = 2,
  toolong = 3.5,
  tsne_perplexity = 15,
  tsne_max_iter = 5000
)
}
\arguments{
\item{ptmtable}{A dataset for post-translational modifications. Formatted with row's names containing PTM names. The column names should be drugs. Values are numeric values that represent how much the PTM has reacted to the drug.}

\item{keeplength}{Only keep clusters of ptms whose size is larger than this parameter. (I.e keeplength = 2 then keep ("AARS", "ARMS", "AGRS") but not ("AARS", "ARMS")); default is 2}

\item{toolong}{A numeric threshold for cluster separation, defaults to 3.5.}

\item{tsne_perplexity}{The perplexity parameter for Rtsne. Please read Rtsne documentation for more information.}

\item{tsne_max_iter}{Number of iterations for Rtsne. Please read Rtsne documentation for more information.}
}
\value{
A list with these data structures at the given index:
\enumerate{
\item{Clusters in all 3 distance metrices as a list.}
\item{A matrix containing values of 0s and 1s depending on if the PTMs are cocluster with other PTMs, rows and columns are unamed.}
\item{A data frame showing the correlation between ptms (as the rows and the columns) with NAs placed along the diagonal.}
}
}
\description{
This function takes the post-translational modification table and runs it through three calculations of distance:
Euclidean Distance, Spearman Dissimilarity (1 - |Spearman Correlation|), and the average of the two of these.
These calculations find the 'distance' between ptms based upon under what conditions they occur.
These matricies are then run through t-SNE in order to put them into a 3-dimensional space. Additionally, the intersection of these 3 clusters is also created.
A correlation table is also produced based on the Spearman Correlation table.
}
\details{
Please note: t-SNE involves an element of randomness; in order to get the same results, set.seed(#) must be called.
}
\examples{
Example_Output <- MakeClusterList(ex_tiny_ptm_table) #Run function
Example_Output[[1]][1:3] #Display data
#Do we want to have one for adj.consensus? Doesn't seem like it'd be very helpful to view.
utils::head(Example_Output[[3]][, c(1,2,3,4,5)]) #Display data
}
