% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCorrelationNetwork.R
\name{MakeCorrelationNetwork}
\alias{MakeCorrelationNetwork}
\title{Make Correlation Network}
\usage{
MakeCorrelationNetwork(adj.consensus.matrix, ptm.correlation.matrix)
}
\arguments{
\item{adj.consensus.matrix}{Adjacency matrix showing PTM co-cluster relationships from all three t-SNE embeddings}

\item{ptm.correlation.matrix}{A data frame showing the correlation between ptms (as the rows and the columns). NAs are placed along the diagonal.}
}
\value{
A list containing the following data structures at the given index:
\enumerate{
\item{The PTM CoCluster Correlation Network as an edgelist.}
\item{The Gene CoCluster Correlation Network as an edgelist.}
\item{A list of nodes that are in the Gene CoCluster Correlation Network.}
}
}
\description{
Make Correlation Network filters a correlation matrix of PTMs by specific PTMs.
It groups the PTM correlation matrix based on the Genes of PTMs.
By summing these submatrices, it also produces a gene by gene cocluster correlation network shows strength of relationships between proteins using the common clusters between the three distance metrics.
}
\examples{
Example_Output <- MakeCorrelationNetwork(ex_adj_consensus, ex_ptm_correlation_matrix)
head(Example_Output[[1]])
head(Example_Output[[2]])

}
