% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessGMEdgefile.R
\name{MakeDBInput}
\alias{MakeDBInput}
\title{Make file for GeneMania input}
\usage{
MakeDBInput(cccn_matrix, file.path.name = "db_nodes.txt")
}
\arguments{
\item{cccn_matrix}{matrix representing the common clusters from the three distance calculations' clusters}

\item{file.path.name}{path for the output file; defaultsto db_nodes.txt}
}
\value{
A file with all of the gene names which can be copy and pasted into the GeneMania cytoscape app, data frame of the names of the genes
}
\description{
This function outputs a file the user can take to create the GeneMania edgefile.
}
\examples{
cccn.cfn.tools:::ex.make_db_input(ex.cccn_matrix)
}
