% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeyFunctions.R
\name{NA.to.zero.func}
\alias{NA.to.zero.func}
\title{Replace NAs with 0}
\usage{
NA.to.zero.func(df)
}
\arguments{
\item{df}{A data frame where NA's are to be replaced with 0s.}
}
\value{
A data frame with NAs replaced by 0.
}
\description{
This function replaces all NA's in a data frame with 0 values.
}
\examples{
NA.to.zero.func(data.frame(a = c(0, 1), b = c(2, 0)))
}
