% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawDataProcessing.R
\name{name.peptide}
\alias{name.peptide}
\title{Create PTM Peptide Names}
\usage{
name.peptide(genes, modification = "p", sites, aa, pepsep = ";")
}
\arguments{
\item{genes}{A character string of one or more gene symbols.}

\item{modification}{A character string indicating modification shorthand.
Defaults to \code{"p"}.}

\item{sites}{A character string of one or more site positions.}

\item{aa}{A character string containing the amino acid code.}

\item{pepsep}{A separator used in \code{genes} and \code{sites} for ambiguous entries.
Defaults to \code{";"}.}
}
\value{
A character string containing unique peptide names.
}
\description{
Builds standardized PTM peptide names from gene symbols, modification type,
amino acid, and site information. Supports ambiguous entries separated by a
delimiter.
}
\examples{
name.peptide("MAPK1; MAPK3", sites = "185; 204", aa = "T")
}
