% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{outersect}
\alias{outersect}
\title{Symmetric set difference of two vectors}
\usage{
outersect(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A sorted vector of elements present in exactly one of \code{x} or \code{y}.
}
\description{
Returns elements that are in either \code{x} or \code{y} but not in both
(i.e., the outer- or exclusive union).
}
