% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{package.PTMsToPathways.patterns}
\alias{package.PTMsToPathways.patterns}
\title{Patterns}
\format{
Character vector
\describe{
\item{Entries}{Unacceptable patterns}
}
}
\source{
"Made by user"
}
\usage{
package.PTMsToPathways.patterns
}
\description{
A list of unacceptable patterns that need to be removed from PTM names
}
\examples{
package.PTMsToPathways.patterns
}
\keyword{internal}
