% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot}
\alias{Plot}
\title{A helper function that plots data from the outputs of various different functions}
\usage{
Plot(dir = "plots", fun = ".")
}
\arguments{
\item{fun}{If you pass in the name of a specific function you want to plot, then it will skip the asking step. A dot is default and will run the asking step}
}
\description{
A helper function that plots data from the outputs of various different functions
}
\examples{
Plot("MakeClusterList")
}
