% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterFilteredNetwork.R
\name{process_correlation_edges}
\alias{process_correlation_edges}
\title{Process Correlation Edges}
\usage{
process_correlation_edges(mode = "lower")
}
\arguments{
\item{mode}{A string specifying the graph mode. Default is "lower".}
}
\value{
A data frame of correlation edges.
}
\description{
This function processes correlation edges from a given correlation matrix
}
\examples{
process_correlation_edges(cor_matrix)
}
