% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGMedgefile.R
\name{processGMedgefile}
\alias{processGMedgefile}
\title{Process GM Edge File}
\usage{
processGMedgefile(
  gm.edgefile.path,
  gm.nodetable.path,
  nodenames,
  gm.network.name = "gm_network"
)
}
\arguments{
\item{gm.edgefile.path}{path to GeneMANIA edgefile}

\item{gm.nodetable.path}{path to GeneMANIA nodetable}
}
\value{
GeneMANIA ppi network table
}
\description{
This function process GM edgefile.
}
\examples{
cccn.cfn.tools::ex.processGMedgefile("path/to/edgefile", "path/to/nodetable", ex.nodenames, ex.gm.network)
}
