% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayCrosstalkNetwork.R
\name{ReadBioplanetFile}
\alias{ReadBioplanetFile}
\title{Read Bioplanet File}
\usage{
ReadBioplanetFile(bioplanet.file = "bioplanet.csv")
}
\arguments{
\item{bioplanet.file}{Path to a delimited Bioplanet text file.
Common delimiters such as comma, tab, semicolon, and pipe are supported.}
}
\value{
A named list of pathways and associated gene symbols.
}
\description{
Reads a Bioplanet CSV file and converts it into a named list of pathways,
where each list element is a character vector of gene symbols.
}
