% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{brca_clusterlist_data}
\alias{brca_clusterlist_data}
\title{BRCA Cluster List Data}
\format{
list containing the 3 outputs of MakeClusterList
\describe{
\item{Common Clusters}{231 lists of PTMs found to cluster, see Common Clusters}
\item{Adjacency Matrix}{4237 x 4237, see Adjacency Consensus Matrix}
\item{PTM correlation matrix}{4237 x 4237, see Correlation Matrix}
}
}
\source{
"MakeClusterList on the BRCA data"
}
\usage{
data(brca_clusterlist_data)
}
\description{
Output of MakeClusterList on the BRCA data
}
\examples{
head("brca_clusterlist_data[[1]]")
head("brca_clusterlist_data[[2]]")
head("brca_clusterlist_data[[3]]")

}
\keyword{internal}
