% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildClusterFilteredNetwork.R
\name{BuildClusterFilteredNetwork}
\alias{BuildClusterFilteredNetwork}
\title{Build PPI Network}
\usage{
BuildClusterFilteredNetwork(
  stringdb.edges,
  genemania.edges,
  kinsub.edges,
  gene.cccn.edges,
  db.filepaths = c()
)
}
\arguments{
\item{stringdb.edges}{Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database}

\item{genemania.edges}{GeneMANIA network of protein-protein interactions from the genes of study; defaults to NA}

\item{kinsub.edges}{TODO}

\item{gene.cccn.edges}{TODO}

\item{db.filepaths}{A vector of paths to the additional ppi network files; defaults to an empty vector}
}
\value{
A list containing the following data structures at the given index:
\strong{1} All the database data.frames stringdb, genemania, and kinsub bound together in a data.frame
\strong{2} A version of ppi.network with only the edges that exist in gene.cccn and have non-zero weights
}
\description{
Uses all inputted database information on protein-protein interactions in order to create a combined PPI-network.
All edges in the PPI network are checked against the cocluster correlation network to ensure existence and a non-zero weight.
If either of these conditions are not met, then it will be removed from the list of PPI edges.
This new, cluster filtered network is then assigned to the global namespace.
}
\examples{
Example_Output <- ClusterFilteredNetwork(ex.stringdb.edges, ex.gm.edges, "need example", "need example", "need example")
utils::head(Example_Output)
}
