% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPPINetwork.R
\name{BuildPPINetwork}
\alias{BuildPPINetwork}
\title{Build PPI Network}
\usage{
BuildPPINetwork(
  string.edges = NA,
  gm.network = NA,
  db.filepaths = c(),
  ppi.network.name = "ppi.network"
)
}
\arguments{
\item{gm.network}{network produced by GeneMANIA if that was used; defaults to NA}

\item{db.filepaths}{vector of filepaths to data from other databases; defaults to empty vector}

\item{ppi.network.name}{desired name of the output ppi network; defaults to "ppi.network"}

\item{cccn.matrix}{dataframe of dataframes that represent the common clusters from the three distance calculations' clusters}
}
\value{
A data frame of combined edges from STRINGdb and provided database entries
}
\description{
Uses all inputted database information on protein-protein interactions in order to create a combined PPI-network usingall inputted data.
}
\examples{
gmpath <- "genemania-interactions.txt"
pack <- "cccn.cfn.tools"
gmfile <- system.file("genemania", gmpath, package = pack, mustWork = TRUE)
cccn.cfn.tools:::ex.FindPPIEdges(ex.cccn.matrix, ppi.network.name = "ex.ppi.network")
}
