% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPPINetwork.R
\name{BuildPPINetwork}
\alias{BuildPPINetwork}
\title{Build PPI Network}
\usage{
BuildPPINetwork(
  stringdb.edges = NA,
  gm.network = NA,
  db.filepaths = c(),
  ppi.network.name = "ppi.network"
)
}
\arguments{
\item{stringdb.edges}{Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database}

\item{gm.network}{GeneMANIA network of protein-protein interactions from the genes of study; defaults to NA}

\item{db.filepaths}{A vector of paths to the additional ppi network files; defaults to an empty vector}

\item{ppi.network.name}{Desired name for the output protein-protein interaction network using all entered database input; defaults to ppi.network}
}
\value{
A dataframe representing how strongly proteins are known to interact based on previous research contained within provided database networks
}
\description{
Uses all inputted database information on protein-protein interactions in order to create a combined PPI-network.
}
\examples{
gmpath <- "genemania-interactions.txt"
pack <- "cccn.cfn.tools"
gmfile <- system.file("genemania", gmpath, package = pack, mustWork = TRUE)
BuildPPINetwork(ex.cccn.matrix, ppi.network.name = "ex.ppi.network")
}
