% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterFilteredNetwork.R
\name{ClusterFilteredNetwork}
\alias{ClusterFilteredNetwork}
\title{ClusterFilteredNetwork}
\usage{
ClusterFilteredNetwork(cccn_matrix, ppi.network, cfn.name = "cfn")
}
\arguments{
\item{accuracy}{How close a PPI can weight be to a correlation weight to accept the edge. It is recommeneded that genemania users do not increase this parameter as it defaults to the maximum sig figs for a genemania egde weight.}
}
\description{
This function finds the intersection between a list of PPI edges and a matrix (which it interprets as a weighted adjacency matrix). Assigns a cfn_network variable to the global environment, which is a list of the common edges between the PPI edges and cccn_matrix.
}
\examples{
cccn.cfn.tools:::ex.ClusterFilteredNetwork()
}
