% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterFilteredNetwork.R
\name{ClusterFilteredNetwork}
\alias{ClusterFilteredNetwork}
\title{ClusterFilteredNetwork}
\usage{
ClusterFilteredNetwork(cccn.matrix, ppi.network, cfn.name = "cfn")
}
\arguments{
\item{cccn.matrix}{Matrix representing the common clusters from the three distance calculations' clusters}

\item{ppi.network}{A data frame of combined edges from STRINGdb and provided database entries}

\item{cfn.name}{Desired name of the output cluster filtered network}
}
\value{
cluster filtered network
}
\description{
This function finds the intersection between a list of PPI edges and a matrix (which it interprets as a weighted adjacency matrix). Assigns a cfn variable to the global environment, which is a list of the common edges between the PPI edges and cccn.matrix.
}
\examples{
ClusterFilteredNetwork(ex.cccn_matrix, ex.ppi_network, cfn.name = "example.cfn")
}
