% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{connectNodes.all}
\alias{connectNodes.all}
\title{Connect All Nodes}
\usage{
connectNodes.all(nodepair, ig.graph = NULL, edgefile, newgraph = FALSE)
}
\arguments{
\item{nodepair}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{ig.graph}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{edgefile}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{newgraph}{PLACEHOLDER PARAMETER DESCRIPTION}
}
\description{
Use all_shortest_paths and returns just the edge file
}
