% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{cytoscape.graph.PCN.pathways}
\alias{cytoscape.graph.PCN.pathways}
\title{Graph Cytoscape PCN Pathways}
\usage{
cytoscape.graph.PCN.pathways(
  PCN = pathway.crosstalk.network,
  net.name,
  Jaccard.edges = TRUE
)
}
\arguments{
\item{PCN}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{net.name}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{Jaccard.edges}{PLACEHOLDER PARAMETER DESCRIPTION}
}
\description{
For graphing Pathway Crosstalk Networks (PCNs) in Cytoscape
}
