% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.bioplanet}
\alias{ex.bioplanet}
\title{Bioplanet Example}
\format{
Dataframe with 100 rows and 4 columns:
\describe{
\item{PATHWAY_ID}{ID of the pathway in bioplanet}
\item{PATHWAY_NAME}{Common name of the pathway}
\item{GENE_ID}{ID of the gene in bioplanet}
\item{GENE_SYMBOL}{Common name of the gene}
}
}
\source{
\url{https://tripod.nih.gov/bioplanet/}
}
\usage{
ex.bioplanet
}
\description{
(description)
}
\examples{
data(ex.bioplanet)
head(ex.bioplanet)
}
\keyword{datasets}
