% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.funckey}
\alias{ex.funckey}
\title{Funckey}
\format{
Dataframe of 24339 rows and 8 columns:
\describe{
\item{Gene.Name}{}
\item{Approved.Name}{}
\item{Hugo.Gene.Family}{}
\item{HPRD.Function}{}
\item{nodeType}{}
\item{Domains}{}
\item{Compartment}{}
\item{Compartment.Overview}{}
}
}
\source{
""
}
\usage{
ex.funckey
}
\description{
DESCRIPTION
}
\examples{
head(ex.funckey)
}
\keyword{internal}
