% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.gm.network}
\alias{ex.gm.network}
\title{Full GeneMANIA Network Example}
\format{
Dataframe of 601,552 rows and 4 columns:
\describe{
\item{Gene.1}{The gene interacting with Gene.2}
\item{Gene.2}{The gene interacting with Gene.1}
\item{raw.weights}{Strength of the interaction}
\item{GM.weights}{Weights from Genemania file}
}
}
\source{
\url{https://apps.cytoscape.org/apps/genemania}
}
\usage{
ex.gm.network
}
\description{
Network of relationships between genes based on information from the GeneMANIA database.
}
\examples{
head(ex.gm.network)
}
\keyword{internal}
