% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.pathways.list}
\alias{ex.pathways.list}
\title{Pathways List Example}
\format{
List of 1657 elements:
\describe{
\item{pathway}{A pathway name with the list of genes involved}
}
}
\source{
"created from a Bioplanet.csv file in PathwayCrosstalkNetwork"
}
\usage{
ex.pathways.list
}
\description{
A named list with names of Pathways and string vectors that hold all the gene symbols from those pathways
}
\examples{
head(ex.pathways.list)
}
\keyword{internal}
