% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.ppi_network}
\alias{ex.ppi_network}
\title{PPI Network Example}
\format{
Dataframe of 1 row and 4 columns:
\describe{
\item{Gene.1}{The gene interacting with Gene.2}
\item{Gene.2}{The gene interacting with Gene.1}
\item{Weight}{Strength of the interaction}
\itm{edgeType}{Type of interaction between Gene.1 and Gene.2}
}
}
\source{
"created by FindPPIEdges"
}
\usage{
ex.ppi_network
}
\description{
Network of relationships between genes based on PPI database STRINGdb.
}
\examples{
data(ex.ppi_network)
head(ex.ppi_network)
}
\keyword{datasets}
