% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Examples.R
\name{ex.processGMedgefile}
\alias{ex.processGMedgefile}
\title{Process GeneMANIA example}
\usage{
ex.processGMedgefile(epath, npath, ex.nodenames, ex.gm.network)
}
\description{
Process GeneMANIA example
}
\keyword{internal}
