% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.ptmtable}
\alias{ex.ptmtable}
\title{PTM Table Example}
\format{
Dataframe of 100 rows and 70 columns:
\describe{
\item{rows}{Post-translational modifications}
\item{columns}{Experimental condition}
\item{values}{Expression as found by mass spectrometry}
}
}
\source{
"manually created"
}
\usage{
ex.ptmtable
}
\description{
Post-translational modification data under certain environmental conditions.
}
\examples{
data(ex.ptmtable)
head(ex.ptmtable)
}
\keyword{datasets}
