% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.ptmtable}
\alias{ex.ptmtable}
\title{PTM Table Example}
\format{
Dataframe of 497 rows and 72 columns:
\describe{
\item{first 3 columns}{Gene, modification, and Amino Acid respectively}
\item{columns}{Names of drugs that were used on PTMs}
\item{row}{Number of PTMs}
\item{values}{Expression as found by mass spectrometry}
}
}
\source{
"manually created"
}
\usage{
ex.ptmtable
}
\description{
Post-translational modification data under certain environmental conditions.
}
\examples{
head(ex.ptmtable)
}
\keyword{internal}
