% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.ptmtable}
\alias{ex.ptmtable}
\title{PTM Table Example}
\format{
Dataframe of 100 rows and 70 columns:
\describe{
\item{}{}
...
}
}
\source{
source info (or \code{"manually created"})
}
\usage{
ex.ptmtable
}
\description{
Post-translational modification data under certain environmental conditions.
}
\examples{
data(ex.ptmtable)
head(ex.ptmtable)
}
\keyword{datasets}
