% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.PTPedgelist}
\alias{ex.PTPedgelist}
\title{PTP Edgelist Example}
\format{
Dataframe of 19 rows and 4 columns:
\describe{
\item{V1}{A pathway}
\item{V2}{A pathway}
\item{jaccard.values}{The jaccard similarity of the two pathways sets of genes}
\item{PTPscore}{The sum of all CPEs for every cluster where both pathways have a nonzero score}
}
}
\source{
"Created in PathwayCrosstalkNetwork"
}
\usage{
ex.PTPedgelist
}
\description{
An edgelist filtered from the combination of every pathway based on nonzero Jaccard & PTP score
}
\examples{
head(ex.PTPedgelist)
}
\keyword{internal}
