% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.stringdb.edges}
\alias{ex.stringdb.edges}
\title{STRINGdb Edges example}
\format{
Dataframe of 205 rows and 4 columns:
\describe{
\item{Gene.1}{First gene interacting}
\item{Gene.2}{Second gene interacting}
\item{Interaction}{The type of interaction}
\item{STRINGdb.combined_score}{The weight of the interaction between these two genes as described by STRINGdb}
}
}
\source{
"created by GetSTRINGdb"
}
\usage{
ex.stringdb.edges
}
\description{
List of edges between genes in dataset.
}
\examples{
head(ex.stringdb.edges)
}
\keyword{internal}
