% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex.tsne.matrices}
\alias{ex.tsne.matrices}
\title{TSNE matrices Example}
\format{
List of 3 Distance metrics (Euclidean, Spearman, and the average of Spearman and Euclidean (SED)) and the clusters that form when they are run through tsne:
\describe{
\item{Euclidean}{A list of 6 clusters of ptms based on their Euclidean distance and positional distance determined by tsne}
\item{Spearman}{A list of 6 clusters of ptms based on their Spearman dissimilarity and positional distance determined by tsne}
\item{SED}{A list of 6 clusters of ptms based on the average of their Euclidean distance and Spearman Dissimilarity and positional distance determined by tsne}
}
}
\source{
"created by MakeClusterList"
}
\usage{
ex.tsne.matrices
}
\description{
List of of lists of dataframes of ptm positions in 3D space to show their relationships with each other based on distance.
}
\examples{
data(ex.tsne.matrices)
head(ex.tsne.matrices)
}
\keyword{datasets}
