% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_pathway_crosstalk_network}
\alias{ex_pathway_crosstalk_network}
\title{Pathway Crosstalk Network Example}
\format{
Matrix with 37 rows and 4 columns:
\describe{
\item{V1}{Pathway One}
\item{V2}{Pathway Two}
\item{Weight}
\item{Interaction}{Jaccard similarity between the two pathways}
}
}
\source{
"created by PathwayCrosstalkNetwork"
}
\usage{
ex_pathway_crosstalk_network
}
\description{
Contains pathway source-target columns and weight, along with the interaction type.
}
\examples{
head(ex_pathway_crosstalk_network)
}
\keyword{internal}
