% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_pathways_list}
\alias{ex_pathways_list}
\title{Pathways List Example}
\format{
List of 9 elements:
\describe{
\item{pathway}{A pathway name with the list of genes involved}
}
}
\source{
"created from a Bioplanet.csv file in PathwayCrosstalkNetwork"
}
\usage{
ex_pathways_list
}
\description{
A named list with names of Pathways and string vectors that hold all the gene symbols from those pathways
}
\examples{
head(ex_pathways_list)
}
\keyword{internal}
