% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ex_PCNedgelist}
\alias{ex_PCNedgelist}
\title{Pathway Crosstalk Network Edgelist Example}
\format{
Matrix with 19 rows and 4 columns:
\describe{
\item{V1}{Pathway One}
\item{V2}{Pathway Two}
\item{Jaccard Weight}{Jaccard Similarity between two pathways}
\item{PTM cluster evidence}{Similarity score based on clusters PTMs share}
}
}
\source{
"created by PathwayCrosstalkNetwork"
}
\usage{
ex_PCNedgelist
}
\description{
Similar to the Pathway Crosstalk Network but uses different weights.
}
\examples{
head(ex_PCNedgelist)
}
\keyword{internal}
