% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{filter.edges.between}
\alias{filter.edges.between}
\title{Filter Edges Inbetween}
\usage{
filter.edges.between(nodes1, nodes2, edge.file, convert = FALSE)
}
\arguments{
\item{nodes1}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{nodes2}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{edge.file}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{convert}{PLACEHOLDER PARAMETER DESCRIPTION}
}
\description{
This function narrows the search only for edges between two sets of nodes
}
