% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCorrelationNetwork.R
\name{FindCommonClusters}
\alias{FindCommonClusters}
\title{Finds common elements between clusters in three lists.}
\usage{
FindCommonClusters(list1, list2, list3, klength)
}
\arguments{
\item{list1}{A list of ptms.}

\item{list2}{A list of ptms.}

\item{klength}{Minimum size of intersections to keep.}
}
\value{
A list of common clusters.
}
\description{
Finds common elements between clusters in three lists.
}
\examples{
FindCommonClusters(list1, list2, list3, keeplength)
}
