% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherPPIData.R
\name{formatKinsubTable}
\alias{formatKinsubTable}
\title{Format Kinsub Table}
\usage{
formatKinsubTable(
  kinasesubstrate.filename = "Kinase_Substrate_Dataset.txt",
  gene.cccn.nodes
)
}
\arguments{
\item{kinasesubstrate.filename}{The path to the kinase substrate database file from https://www.phosphosite.org/staticDownloads}

\item{gene.cccn.nodes}{A list of nodes that are in the Gene CoCluster Correlation Network derived from common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))}
}
\value{
An edgelist filtered by the gene cccn and nodenames
}
\description{
Include kinase substrate dataset from PhosphoSitePlus https://www.phosphosite.org/staticDownloads
}
