% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{get.co.clustered.ptms}
\alias{get.co.clustered.ptms}
\title{Get CoClustered PTMs}
\usage{
get.co.clustered.ptms(edge.file)
}
\arguments{
\item{edge.file}{}
}
\description{
This function takes an edge file, retrieves only co-clustered PTM CCCN edges and links them to their gene nodes, returning an edge file
}
