% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindPPIEdges.R
\name{get.GM.edgefile}
\alias{get.GM.edgefile}
\title{Loads and filters the GeneMania file given a vector of gene names.Add commentMore actions}
\usage{
get.GM.edgefile(nodenames, gmfilepath = "genemania-interactions.txt")
}
\arguments{
\item{nodenames}{A vector containing the names of the relevant genes.}

\item{gmfilepath}{The path to the GeneMania file initialized to "genemania-interactions.txt".}
}
\value{
A data frame with the relevant GeneMania data.
}
\description{
This helper function loads the GeneMania file
and filters out required nodes.
}
