% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherPPIData.R
\name{GetGeneMANIA.edges}
\alias{GetGeneMANIA.edges}
\title{Get GeneMANIA Edges}
\usage{
GetGeneMANIA.edges(gm.results.path, gene.cccn.nodes)
}
\arguments{
\item{gm.results.path}{Path to GeneMANIA results text file}

\item{gene.cccn.nodes}{A list of nodes that are in the Gene CoCluster Correlation Network derived from common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))}
}
\value{
Data frame of consisting of the network of interactions from the genes of study
}
\description{
This function processes the GM results text file and returns it as a filtered edgelist.
}
\details{
To get the GeneMANIA results text file, click on the three lines in the upper right corner.
This should be under the GeneMANIA side window beside the species.
Click "Export Results". The path to this file is the gm.results.path.
An visual depiction of these instructions in the vignette PTMsToPathways under the GeneMANIA section.
}
\examples{
ex.gm.results.path <- system.file("extdata/ex_gm_edgetable.csv", package = "PTMsToPathways")
example.GeneMANIA.edges <- GetGeneMANIA.edges(ex.gm.results.path, ex.gene.cccn.nodes)
utils::head(example.GeneMANIA.edges)
}
