% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherPPIData.R
\name{GetSTRINGdb.edges}
\alias{GetSTRINGdb.edges}
\title{Get STRINGdb PPI data}
\usage{
GetSTRINGdb.edges(gene.cccn.edges, gene.cccn.nodes)
}
\arguments{
\item{gene.cccn.edges}{A dataframe showing interactions relationships between proteins using common PTM clusters derived from three distance metrics (Euclidean, Spearman, and Combined (SED))}

\item{gene.cccn.nodes}{A list of nodes that are in the Gene CoCluster Correlation Network derived from common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))}
}
\value{
Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database and a list of gene names
}
\description{
This function finds protein-protein interaction weights by consulting utilizing the STRINGdb database.
}
\details{
The full example takes ~10 minutes to load, so it has been commented out and the results are displayed.
}
\examples{
# GetSTRINGdb(ex.gene.cccn)
utils::head(ex.stringdb.edges)
utils::head(ex.nodenames)
}
