% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherPPIData.R
\name{GetSTRINGdb}
\alias{GetSTRINGdb}
\title{Get STRINGdb}
\usage{
GetSTRINGdb(
  cccn.matrix,
  STRINGdb.name = "string.edges",
  nodenames.name = "nodenames"
)
}
\arguments{
\item{cccn.matrix}{dataframe of dataframes that represent the common clusters from the three distance calculations' clusters}

\item{STRINGdb.name}{desired name for the output STRINGdb data frame; defaults to "string.edges"}

\item{nodes.name}{desired name for the output containing the list of all of the gene names; defaults to "nodenames"}
}
\value{
A data frame of edges from STRINGdb showing interactions between genes
}
\description{
This function finds protein-protein interaction edges by consulting utilizing the STRINGdb database.
}
\examples{
cccn.cfn.tools:::ex.FindPPIEdges(ex.cccn.matrix, "ex.string.edges", "ex.nodenames")
}
