% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherPPIData.R
\name{GetSTRINGdb}
\alias{GetSTRINGdb}
\title{Get STRINGdb PPI data}
\usage{
GetSTRINGdb(
  cccn.matrix,
  STRINGdb.name = "string.edges",
  nodenames.name = "nodenames"
)
}
\arguments{
\item{cccn.matrix}{A matrix showing strength of relationships between proteins using common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))}

\item{STRINGdb.name}{Desired name for the output STRINGdb data frame; defaults to "string.edges"}

\item{nodes.name}{Desired name for list of gene names; defaults to nodenames}
}
\value{
Data frame of consisting of the network of interactions from the genes of study pulled from the STRINGdb database and a list of gene names
}
\description{
This function finds protein-protein interaction weights by consulting utilizing the STRINGdb database.
}
\examples{
cccn.cfn.tools:::ex.FindPPIEdges(ex.cccn.matrix, "ex.string.edges", "ex.nodenames")
}
