% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{GraphCfn}
\alias{GraphCfn}
\title{Graph Cluster Filtered Network
This function wraps RCy3 graphing in Cytoscape and sets node and edge visual properties}
\usage{
GraphCfn(
  cfn.edges,
  cfn.nodes,
  Network.title = "CFN",
  Network.collection = "PTMsToPathways",
  visual.style.name = "PTMsToPathways.style"
)
}
\arguments{
\item{cfn.edges}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{cfn.nodes}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{Network.title}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{Network.collection}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{visual.style.name}{PLACEHOLDER PARAMETER DESCRIPTION}
}
\description{
Graph Cluster Filtered Network
This function wraps RCy3 graphing in Cytoscape and sets node and edge visual properties
}
