% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphing.R
\name{GraphCfn}
\alias{GraphCfn}
\title{Graph Cluster Filtered Network}
\usage{
GraphCfn(cfn, Network.title = "cfn", Network.collection = "cccn.cfn.tools")
}
\arguments{
\item{cfn}{A version of ppi.network with only the edges that exist in cccn.matrix and have non-zero weights}

\item{Network.title}{Desired title for the created Cytoscape Network}

\item{Network.collection}{Desired name for the collection created on Cytoscape in which the network will reside}
}
\value{
A cytoscape graph of the cluster filtered network
}
\description{
Creates a cytoscape graph of the cluster filtered network. Ensure that you have the cytoscape app open and the RCy3 package downloaded and libraried.
}
\examples{
GraphCFN(ex.cfn)
}
