% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphing.R
\name{GraphCfn}
\alias{GraphCfn}
\title{Graph Cluster Filtered Network}
\usage{
GraphCfn(cfn)
}
\arguments{
\item{cfn}{A version of ppi.network with only the edges that exist in cccn.matrix and have non-zero weights}
}
\value{
A cytoscape graph of the cluster filtered network
}
\description{
Creates a cytoscape graph of the cluster filtered network.
}
\examples{
gmpath <- "genemania-interactions.txt"
pack <- "cccn.cfn.tools"
gmfile <- system.file("genemania", gmpath, package = pack, mustWork = TRUE)
cccn.cfn.tools:::ex.FindPPIEdges(ex.cccn.matrix, ppi.network.name = "ex.ppi.network")
}
