% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphing.R
\name{GraphCfn}
\alias{GraphCfn}
\title{Graph Cluster Filtered Network}
\usage{
GraphCfn(cfn, network.name = "cfn.cytoscape")
}
\arguments{
\item{cfn}{A version of ppi.network with only the edges that exist in cccn.matrix and have non-zero weights}
}
\value{
A cytoscape graph of the cluster filtered network
}
\description{
Creates a cytoscape graph of the cluster filtered network. Ensure that you have the cytoscape app open and the RCy3 package downloaded and libraried.
}
\examples{
GraphCFN(ex.cfn)
}
