% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphing.R
\name{GraphCfn}
\alias{GraphCfn}
\title{Graph Cluster Filtered Network}
\usage{
GraphCfn(
  cfn,
  ptmtable,
  funckey = PTMsToPathways::ex.funckey,
  Network.title = "cfn",
  Network.collection = "PTMsToPathways",
  visual.style.name = "PTMsToPathways.style",
  background.color = "#949494",
  edge.label.color = "#17202a",
  node.label.color = "#000000",
  default.font = "Times New Roman",
  node.font.size = 20,
  edge.font.size = 8,
  edge.line.style = "SOLID",
  edge.opacity = 175,
  edge.label.opacity = 255,
  border.opacity = 255,
  node.label.opacity = 255,
  node.fill.opacity = 255
)
}
\arguments{
\item{cfn}{A version of ppi.network with only the edges that exist in cccn.matrix and have non-zero weights}

\item{ptmtable}{A dataset for post-translational modifications. Formatted with numbered rows, and the first column containing PTM names. The rest of the column names should be drugs. Values are numeric values that represent how much the PTM has reacted to the drug.}

\item{funckey}{A table graphing gene names to type of protein; defaults to internal database at PTMsToPathways::ex.funckey}

\item{Network.title}{Desired title for the created Cytoscape Network; defaults to "cfn"}

\item{Network.collection}{Desired name for the collection created on Cytoscape in which the network will reside; defaults to "PTMsToPathways"}

\item{visual.style.name}{Desired name for the visual style created on Cytoscape; defaults to "PTMsToPathways.style"}

\item{background.color}{Hex code of background color of graph; defaults to '#fcf3cf'}

\item{edge.label.color}{Hex code of edge label color of graph; defaults to '#17202a'}

\item{node.label.color}{Hex code of node label color of graph; defaults to '#145a32'}

\item{default.font}{Font style of edge and node names; defaults to "Times New Roman"}

\item{node.font.size}{Font size of the node name; defaults to 12}

\item{edge.font.size}{Font size of the edge name; defaults to 8}

\item{edge.line.style}{Type of edge style; defaults to "SOLID"; options include: "PARALLEL_LINES", "MARQUEE_EQUAL", "DOT", "EQUAL_DASH", "LONG_DASH", "CONTIGUOUS_ARROW", "MARQUEE_DASH", "DASH_DOT", "BACKWARD_SLASH", "FORWARD_SLASH", "VERTICAL_SLASH", "SOLID", "SEPARATE_ARROW", "MARQUEE_DASH_DOT", "ZIGZAG", "SINEWAVE"}

\item{edge.opacity}{Opacity of the edge line on a scale of 0 - 255 with 0 being transparent; defaults to 175}

\item{edge.label.opacity}{Opacity of the edge label on a scale of 0 - 255 with 0 being transparent; defaults to 255}

\item{border.opacity}{Opacity of the node border on a scale of 0 - 255 with 0 being transparent; defaults to 255}

\item{node.label.opacity}{Opacity of the node label on a scale of 0 - 255 with 0 being transparent; defaults to 255}

\item{node.fill.opacity}{Opacity of the node fill on a scale of 0 - 255 with 0 being transparent; defaults to 255}
}
\value{
A cytoscape graph of the cluster filtered network
}
\description{
Creates a cytoscape graph of the cluster filtered network. Ensure that you have the Cytoscape app open and the RCy3 package downloaded.
RCy3 is required for this function. To download, run:
if (!require("BiocManager", quietly = TRUE)) install.packages("BiocManager")
BiocManager::install("RCy3")
If RCy3 is installed and the Cytoscape App is open and an error still occurs, run cytoscapePing() to ensure connection to the Cytoscape Interface and try again.
}
\details{
Graph Key
\itemize{
\item Node Size
\itemize{
\item Greater the node size, larger the absolute value of the score
}
\item Blue Node
\itemize{
\item Negative score
}
\item Yellow Node
\itemize{
\item Positive score
}
\item Green Node
\itemize{
\item Approximately zero score
Node Shapes:
}
\item "ELLIPSE"
\itemize{
\item unknown
}
\item "ROUND_RECTANGLE"
\itemize{
\item receptor tyrosine kinase
}
\item "VEE"
\itemize{
\item SH2 protein
or
\item SH2-SH3 protein
}
\item "TRIANGLE"
\itemize{
\item SH3 protein
}
\item "HEXAGON"
\itemize{
\item tyrosine kinase
}
\item "DIAMOND"
\itemize{
\item SRC-family kinase
}
\item "OCTAGON"
\itemize{
\item kinase
or
\item phosphatase
}
\item "PARALLELOGRAM"
\itemize{
\item transcription factor
}
\item "RECTANGLE"
\itemize{
\item RNA binding protein
Node Border Colors:
}
\item Orange
\itemize{
\item deacetylase
or
\item acetyltransferase
}
\item Blue
\itemize{
\item demethylase
or
\item methyltransferase
}
\item Royal Purple
\itemize{
\item membrane protein
}
\item Red
\itemize{
\item kinase
or
\item tyrosine kinase
or
\item SRC-family kinase
}
\item Yellow
\itemize{
\item phosphatase
or
\item tyrosine phosphatase
}
\item Lilac
\itemize{
\item G protein-coupled receptor
or
\item receptor tyrosine kinase
}
\item Grey
\itemize{
\item default
}
\item Edge Thickness
\itemize{
\item Stronger correlation
Edge Colors:
}
\item Red
\itemize{
\item Phosphorylation
or
\item pp
or
\item controls-phosphorylation-of
}
\item Bright Magenta
\itemize{
\item controls-expression-of
}
\item Dull Magenta
\itemize{
\item controls-transport-of
}
\item Purple
\itemize{
\item controls-state-change-of
}
\item Blood Orange
\itemize{
\item Acetylation
}
\item Lime Green
\itemize{
\item Phystical interactions
}
\item Green
\itemize{
\item BioPlex
}
\item Dull Green
\itemize{
\item in-complex-with
}
\item Seafoam Green
\itemize{
\item experiments
or
\item experiments_transferred
}
\item Cyan
\itemize{
\item database
or
\item database_transferred
}
\item Teal
\itemize{
\item Pathway
or
\item Predicted
}
\item Dark Turquoise
\itemize{
\item Genetic interactions
}
\item Yellow-Orange
\itemize{
\item correlation
}
\item Royal Blue
\itemize{
\item negative correlation
}
\item Bright Yellow
\itemize{
\item positive correlation
}
\item Grey
\itemize{
\item combined_score
}
\item Dark Grey
\itemize{
\item merged
}
\item Light Grey
\itemize{
\item intersect
}
\item Black
\itemize{
\item peptide
}
\item Orange
\itemize{
\item homology
}
\item Dull Orange
\itemize{
\item Shared protein domains
}
\item White
\itemize{
\item Default
Arrow Types:
}
\item Arrow
\itemize{
\item Phosphorylation
or
\item pp
or
\item controls-phosphorylation-of
or
\item controls-expression-of
or
\item controls-transport-of
or
\item controls-state-change-of
or
\item Acetylation
}
\item No Arrow
\itemize{
\item Default
}
}
}
\examples{
# GraphCFN(ex.cfn)
# See vignette for default graph
}
