% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{make.cytoscape.node.file}
\alias{make.cytoscape.node.file}
\title{Make Cytoscape Node File}
\usage{
make.cytoscape.node.file(
  edge.file,
  funckey,
  ptmtable,
  include.gene.data = FALSE,
  include.coclustered.PTMs = FALSE
)
}
\arguments{
\item{edge.file}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{funckey}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{ptmtable}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{include.gene.data}{PLACEHOLDER PARAMETER DESCRIPTION}

\item{include.coclustered.PTMs}{PLACEHOLDER PARAMETER DESCRIPTION}
}
\description{
Makes a nodefile to be entered into Cytoscape
}
