% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphing.R,
%   R/CytoscapeGraphingFunctions_Named.r
\name{make.genepep.edges}
\alias{make.genepep.edges}
\title{Title}
\usage{
make.genepep.edges(peptide.edgefile)

make.genepep.edges(peptide.edgefile)
}
\arguments{
\item{peptide.edgefile}{}
}
\description{
Title

Title
}
