% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeClusterList.R
\name{MakeClusterList}
\alias{MakeClusterList}
\title{Populates the global environment with cluster groupings based on t-SNE data}
\usage{
MakeClusterList(
  ptmtable,
  correlation.matrix.name = "ptmcorrelation.matrix",
  eu.tsne.name = "eu_ptms_list",
  sp.tsne.name = "sp_ptms_list",
  sed.tsne.name = "sed_ptms_list",
  toolong = 3.5
)
}
\arguments{
\item{ptmtable}{A dataset for post-translational modifications.}

\item{toolong}{A numeric threshold for cluster separation, defaults to 3.5.}
}
\description{
This function groups t-SNE data points into clusters using a specified threshold
and visualizes the clusters.
}
\examples{
cccn.cfn.tools:::ex.MakeClusterList(ex.ptmtable, toolong =  3.5)
}
