% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeClusterList.R
\name{MakeClusterList}
\alias{MakeClusterList}
\title{Build Clusters Based on PTM Data Under Like Conditions}
\usage{
MakeClusterList(
  ptmtable,
  correlation.matrix.name = "ptm.correlation.matrix",
  list.name = "clusters.list",
  toolong = 3.5
)
}
\arguments{
\item{ptmtable}{A dataset for post-translational modifications. Formatted with numbered rows, and the first column containing PTM names. The rest of the column names should be drugs. Values are numeric values that represent how much the PTM has reacted to the drug.}

\item{correlation.matrix.name}{Desired name for the correlation matrix to be saved as; defaults to ptm.correlation.matrix}

\item{list.name}{Desired name for the lists of clusters to be saved as; defaults to clusters.list}

\item{toolong}{A numeric threshold for cluster separation, defaults to 3.5.}
}
\value{
The correlation matrix: A data frame showing the correlation between ptms (as the rows and the columns) with NAs placed along the diagonal; and A list of three-dimensional data frames used to represent ptms in space to show relationships between them based on distances. Based on Euclidean Distance, Spearman Dissimilarity, and SED (the average between the two)
}
\description{
This function takes the post-translational modification table and runs it through three calculations of distance:
Euclidean Distance, Spearman Dissimilarity (1 - |Spearman Correlation|), and the average of the two of these.
These calculations find the 'distance' between ptms based upon under what conditions they occur.
These matricies are then run through t-SNE in order to put them into a 3-dimensional space.
A correlation table is also produced based on the Spearman Correlation table.
}
\details{
Please note: t-SNE involves an element of randomness; in order to get the same results, set.seed(#) must be called.
}
\examples{
MakeClusterList(ex.ptmtable, "ex.ptm.correlation.matrix", "ex.clusters.list", 3.5)
utils::head(ex.ptm.correlation.matrix)
utils::head(ex.clusters.list[[1]])
utils::head(ex.clusters.list[[2]])
utils::head(ex.clusters.list[[3]])
}
