% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeClusterList.R
\name{MakeClusterList}
\alias{MakeClusterList}
\title{Populates the global environment with cluster groupings based on t-SNE data}
\usage{
MakeClusterList(
  ptmtable,
  correlation.matrix.name = "ptm.correlation.matrix",
  list.name = "tsne.matrices",
  toolong = 3.5
)
}
\arguments{
\item{ptmtable}{A dataset for post-translational modifications.}

\item{correlation.matrix.name}{The desired name for the PTM correlation matrix created}

\item{list.name}{The desired name for the data structure storing the 3 t-SNE matrices}

\item{toolong}{A numeric threshold for cluster separation, defaults to 3.5.}
}
\description{
This function groups t-SNE data points into clusters using a specified threshold
and visualizes the clusters.
}
\examples{
MakeClusterList(ex.ptmtable, correlation.matrix.name = "example.cor", list.name = "example.tsne", toolong =  3.5)
}
