% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCorrelationNetwork.R
\name{MakeCorrelationNetwork}
\alias{MakeCorrelationNetwork}
\title{Make Correlation Network}
\usage{
MakeCorrelationNetwork(common.clusters, ptm.correlation.matrix)
}
\arguments{
\item{common.clusters}{A list of clusters. Ideally the ones found by MakeClusterList in common_clusters}

\item{ptm.correlation.matrix}{A data frame showing the correlation between ptms (as the rows and the columns). NAs are placed along the diagonal.}
}
\value{
The list of common clusters between all three distance metrics (Euclidean, Spearman, and SED) and a matrix showing strength of relationships between proteins using the common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))
}
\description{
Make Correlation Network filters a correlation matrix of PTMs by specific PTMs.
It groups the PTM correlation matrix based on the Genes of PTMs.
By summing these submatrices, it also produces a gene by gene cocluster correlation network shows strength of relationships between proteins using the common clusters between the three distance metrics.
}
\examples{
ex.ptm.cor <- ex.ptm.correlation.matrix
MakeCorrelationNetwork(ex.common.clusters, ex.ptm.cor, "ex.ptm.cccn", "ex.gene.cccn")
ex.ptm.cccn[1:5, 1:5]
ex.gene.cccn[1:5, 1:5]

}
