% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCorrelationNetwork.R
\name{MakeCorrelationNetwork}
\alias{MakeCorrelationNetwork}
\title{Make Correlation Network}
\usage{
MakeCorrelationNetwork(adj.consensus, ptm.correlation.matrix)
}
\arguments{
\item{ptm.correlation.matrix}{A data frame showing the correlation between ptms (as the rows and the columns). NAs are placed along the diagonal.}

\item{common.clusters}{A list of clusters. Ideally the ones found by MakeClusterList in common_clusters}
}
\value{
A list containing the following data structures at the given index: \cr
\strong{1} (PTM CoCluster Correlation Network) A matrix that shows the strength of the relationships between PTMs using the common clusters between the three distance metrics \cr
\strong{2} (Gene CoCluster Correlation Network) Similar to \strong{1} but at the level of genes. \cr
\strong{3} The PTM CoCluster Correlation Network as an igraph edgelist object. \cr
\strong{4} The gene CoCluster Correlation Network as an igraph edgelist object.
}
\description{
Make Correlation Network filters a correlation matrix of PTMs by specific PTMs.
It groups the PTM correlation matrix based on the Genes of PTMs.
By summing these submatrices, it also produces a gene by gene cocluster correlation network shows strength of relationships between proteins using the common clusters between the three distance metrics.
}
\examples{
Example_Output <- MakeCorrelationNetwork(ex_adj_consensus, ex.common.clusters)
ex.ptm.cccn[1:5, 1:5]
ex.gene.cccn[1:5, 1:5]

}
