% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCorrelationNetwork.R
\name{MakeCorrelationNetwork}
\alias{MakeCorrelationNetwork}
\title{Make Correlation Network}
\usage{
MakeCorrelationNetwork(
  common.clusters,
  ptm.correlation.matrix,
  ptm.cccn.name = "ptm.cccn",
  gene.cccn.name = "gene.cccn"
)
}
\arguments{
\item{common.clusters}{A list of clusters. Ideally the ones found by MakeClusterList in common_clusters}

\item{ptm.correlation.matrix}{A data frame showing the correlation between ptms (as the rows and the columns). NAs are placed along the diagonal.}

\item{ptm.cccn.name}{The PTM correlation matrix filtered by PTMs that cocluster}

\item{gene.cccn.name}{Desired name for the cocluster correlation network; defaults to gene.cccn}
}
\value{
The list of common clusters between all three distance metrics (Euclidean, Spearman, and SED) and a matrix showing strength of relationships between proteins using the common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))
}
\description{
Make Correlation Network first finds the intersection between the Euclidean, Spearman, and SED cluster matrices in order to find
the intersection between the three groups. It then adds the Genes in these PTMs to a list of common clusters and turns it into an adjacency matrix.
This adjacency matrix is used to filter relevant data --- clusters --- from the Spearman correlation matrix. The resultant
cocluster correlation network shows strength of relationships between proteins using the common clusters between the three distance metrics.
}
\examples{
ex.ptm.cor <- ex.ptm.correlation.matrix
MakeCorrelationNetwork(ex.clusters.list, ex.ptm.cor, 1, "ex.clusters.common", "ex.ptm.cccn", "ex.gene.cccn")
print(ex.clusters.common[c(1, 2, 3)])
utils::head(ex.gene.cccn[, c(1,2,3,4,5)])
}
