% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCorrelationNetwork.R
\name{MakeCorrelationNetwork}
\alias{MakeCorrelationNetwork}
\title{Make Correlation Network}
\usage{
MakeCorrelationNetwork(keeplength = 2)
}
\arguments{
\item{eu_ptms_list}{Matrix containing Euclidean t-SNE coords}

\item{sp_ptms_list}{Matrix containing Spearman t-SNE coords}

\item{sed_ptms_list}{Matrix containing combined t-SNE coords}
}
\value{
An igraph object representing the correlation network.
}
\description{
This functions runs all the steps required for make correlation network
}
\examples{
MakeCorrelationNetwork(eu_ptms_list, sp_ptms_list, sed_ptms_list, keeplength)
}
