% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCorrelationNetwork.R
\name{MakeCorrelationNetwork}
\alias{MakeCorrelationNetwork}
\title{Make Correlation Network}
\usage{
MakeCorrelationNetwork(
  clusterlist,
  ptm.correlation.matrix,
  keeplength = 2,
  clusters.name = "clusters.common",
  filtered.ptm.cor.name = "filtered.ptm.cor",
  cccn.name = "cccn.matrix"
)
}
\arguments{
\item{clusterlist}{A list of three-dimensional data frames used to represent ptms in space to show relationships between them based on distances. Based on Euclidean Distance, Spearman Dissimilarity, and SED (the average between the two)}

\item{ptm.correlation.matrix}{A data frame showing the correlation between ptms (as the rows and the columns). NAs are placed along the diagonal.}

\item{keeplength}{Only keep clusters of ptms whose size is larger than this parameter. (I.e keeplength = 2 then keep ("AARS", "ARMS", "AGRS") but not ("AARS", "ARMS"))}

\item{clusters.name}{Desired name for the common clusters output; defaults to common.clusters}

\item{filtered.ptm.cor.name}{The PTM correlation matrix filtered by PTMs that cocluster}

\item{cccn.name}{Desired name for the cocluster correlation network; defaults to cccn.matrix}
}
\value{
The list of common clusters between all three distance metrics (Euclidean, Spearman, and SED) and a matrix showing strength of relationships between proteins using the common clusters between the three distance metrics (Euclidean, Spearman, and Combined (SED))
}
\description{
Make Correlation Network first finds the intersection between the Euclidean, Spearman, and SED cluster matrices in order to find
the intersection between the three groups. It then adds the Genes in these PTMs to a list of common clusters and turns it into an adjacency matrix.
This adjacency matrix is used to filter relevant data --- clusters --- from the Spearman correlation matrix. The resultant
cocluster correlation network shows strength of relationships between proteins using the common clusters between the three distance metrics.
}
\examples{
ex.ptm.cor <- ex.ptm.correlation.matrix
MakeCorrelationNetwork(ex.clusters.list, ex.ptm.cor, 1, "ex.clusters.common", "ex.filtered.cor.matrix", "ex.cccn.matrix")
print(ex.clusters.common[c(1, 2, 3)])
utils::head(ex.cccn.matrix[, c(1,2,3,4,5)])
}
