% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherPPIData.R
\name{MakeDBInput}
\alias{MakeDBInput}
\title{Make Database Input File}
\usage{
MakeDBInput(cccn.matrix, file.path.name = "db_nodes.txt")
}
\arguments{
\item{cccn.matrix}{matrix representing the common clusters from the three distance calculations' clusters}

\item{file.path.name}{path for the output file; defaults to db_nodes.txt}
}
\value{
A file with all of the gene names which can be copy and pasted into the GeneMania cytoscape app, data frame of the names of the genes
}
\description{
This function outputs a file consisting entirely of gene names, each produced on a new line. This data can be copy and pasted into
a database input in order to get protein-protein interaction data.
}
\examples{
cccn.cfn.tools:::ex.MakeDBInput(ex.nodenames)
}
