% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawDataProcessing.R
\name{merge2cols}
\alias{merge2cols}
\title{Merge Technical Replicate Values}
\usage{
merge2cols(colv1, colv2)
}
\arguments{
\item{colv1}{A numeric value from replicate 1.}

\item{colv2}{A numeric value from replicate 2.}
}
\value{
A numeric value or \code{NA}.
}
\description{
Merges two numeric values from technical replicates by returning \code{NA} when
both are missing, returning the observed value when one is missing, or
returning their mean when both are present.
}
\examples{
merge2cols(10, 14)
merge2cols(NA, 14)
}
