% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{mergeEdges}
\alias{mergeEdges}
\title{Merge Edges}
\usage{
mergeEdges(edgefile)
}
\arguments{
\item{edgefile}{}
}
\description{
Function to merge edges and declutter the Cytoscape network
}
