% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{NodeEdgeKey}
\alias{NodeEdgeKey}
\title{NodeEdgeKey}
\usage{
NodeEdgeKey(visual.style.name = "PTMsToPathways.style")
}
\arguments{
\item{visual.style.name}{Name of the visual style that will be saved in the Cytoscape Interface}
}
\description{
Cytoscape legend for styles
}
