% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayCrosstalkNetwork.R
\name{PathwayCrosstalkNetwork}
\alias{PathwayCrosstalkNetwork}
\title{Converts Bioplanet pathways from (\url{https://tripod.nih.gov/bioplanet/})  into a list of pathways whose elements are the genes in each pathway. Edge weights are either the PTM Cluster Weight or according to the Jaccard Similarity.}
\usage{
PathwayCrosstalkNetwork(file, clusterlist, PCNname = "pcn_matrix")
}
\arguments{
\item{file}{The name of the bioplanet pathway .csv file}

\item{clusterlist}{The list of coclusters made in MakeCorrelationNetwork}
}
\value{
No clue (The result is a matrix with pathway names in columns and individual clusters as rows.)
}
\description{
Converts Bioplanet pathways from (\url{https://tripod.nih.gov/bioplanet/})  into a list of pathways whose elements are the genes in each pathway. Edge weights are either the PTM Cluster Weight or according to the Jaccard Similarity.
}
\examples{
print("TO DO")
}
