% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PathwayCrosstalkNetwork.R
\name{PathwayCrosstalkNetwork}
\alias{PathwayCrosstalkNetwork}
\title{Pathway Crosstalk Network}
\usage{
PathwayCrosstalkNetwork(file = "bioplanet.csv", clusterlist)
}
\arguments{
\item{file}{Either the name of the bioplanet pathway .csv file OR the name of a dataframe loaded in environment, users should only pass in "yourfilename.csv"}

\item{clusterlist}{The list of coclusters made in MakeCorrelationNetwork}
}
\description{
Converts Bioplanet pathways from (\url{https://tripod.nih.gov/bioplanet/})  into a list of pathways whose elements are the genes in each pathway. Edge weights are either the PTM Cluster Weight or according to the Jaccard Similarity.
}
\examples{
PathwayCrosstalkNetwork(ex.bioplanet, ex.common.clusters)
}
