% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterFilteredNetwork.R
\name{process_correlation_edges}
\alias{process_correlation_edges}
\title{Process Correlation Edges}
\usage{
process_correlation_edges(pBound = 0.5, nBound = -0.5, mode = "lower")
}
\arguments{
\item{pBound}{The minimum value an edge weight can be to make it a positive correlation. Default = 0.5}

\item{nBound}{The maximum value an edge weight can be to make it a negative correlation. Default = -0.5}

\item{mode}{A string specifying the graph mode. Default is "lower".}
}
\value{
A data frame of correlation edges.
}
\description{
This function processes correlation edges from a given correlation matrix
}
\examples{
process_correlation_edges(cor_matrix)
}
