% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatherPPIData.R
\name{ProcessGMEdgefile}
\alias{ProcessGMEdgefile}
\title{Process GM Edge File}
\usage{
ProcessGMEdgefile(
  gm.edgefile.path,
  gm.nodetable.path,
  db_nodes.path,
  gm.network.name = "gm.network"
)
}
\arguments{
\item{gm.edgefile.path}{Path to GeneMANIA edgefile}

\item{gm.nodetable.path}{Path to GeneMANIA nodetable}

\item{db_nodes.path}{Path to the node file from MakeDBInput}

\item{gm.network.name}{Desired name for the output genemania network; defaults to gm.network}
}
\value{
Data frame of consisting of the network of interactions from the genes of study
}
\description{
This function processes the GM edgefile and translates it back into gene names using the nodetable.
}
\examples{
ex.edgefile <- system.file("extdata/ex_gm_edgetable.csv", package = "PTMsToPathways")
ex.nodefile <- system.file("extdata/ex_gm_nodetable.csv", package = "PTMsToPathways")
ex.db.nodes  <- system.file("extdata/ex_db_nodes.txt", package = "PTMsToPathways")
ProcessGMEdgefile(ex.edgefile, ex.nodefile, ex.db.nodes, "ex.gm.network")
utils::head(ex.gm.network)
}
