% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processGMedgefile.R
\name{ProcessGMEdgefile}
\alias{ProcessGMEdgefile}
\title{Process GM Edge File}
\usage{
ProcessGMEdgefile(
  gm.edgefile.path,
  gm.nodetable.path,
  db_nodes.path,
  gm.network.name = "gm.network"
)
}
\arguments{
\item{gm.edgefile.path}{path to GeneMANIA edgefile}

\item{gm.nodetable.path}{path to GeneMANIA nodetable}

\item{db_nodes.path}{path to the node file from MakeDBInput}

\item{gm.network.name}{desired name for the output genemania network; defaulted to gm.network}
}
\value{
GeneMANIA ppi network table
}
\description{
This function process GM edgefile.
}
\examples{
cccn.cfn.tools:::ex.ProcessGMEdgefile("edgefile.csv", "nodetable.csv", ex.nodenames, ex.gm.network)
}
