% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphing.R,
%   R/CytoscapeGraphingFunctions_Named.r
\name{remove.autophos}
\alias{remove.autophos}
\title{Title}
\usage{
remove.autophos(edgefile)

remove.autophos(edgefile)
}
\arguments{
\item{edgefile}{}
}
\description{
Title

Title
}
