% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{setEdgeWidths}
\alias{setEdgeWidths}
\title{Set Edge Widths}
\usage{
setEdgeWidths(ffactor = -1.2, log = TRUE)
}
\arguments{
\item{ffactor}{}

\item{log}{}
}
\description{
This helper function will make wider edges if they are two thin or narrow if to thick
}
