% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{setNodeColorToRatios}
\alias{setNodeColorToRatios}
\title{Set Node Color Ratios
Function to sent node size and color to match ratio data in the Cytoscape node table.}
\usage{
setNodeColorToRatios(plotcol)
}
\arguments{
\item{plotcol}{PLACEHOLDER PARAMETER DESCRIPTION}
}
\description{
Set Node Color Ratios
Function to sent node size and color to match ratio data in the Cytoscape node table.
}
