% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{setNodeColorToRowz}
\alias{setNodeColorToRowz}
\title{Set Node Color Rows
This function works well with node data that are normalized by row z-scores}
\usage{
setNodeColorToRowz(plotcol)
}
\arguments{
\item{plotcol}{PLACEHOLDER PARAMETER DESCRIPTION}
}
\description{
Set Node Color Rows
This function works well with node data that are normalized by row z-scores
}
