% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{setNodeMapping}
\alias{setNodeMapping}
\title{Set Node Mapping
Function to set shape and border color according to node type}
\usage{
setNodeMapping(cf = RCy3::getTableColumns("node"))
}
\arguments{
\item{cf}{PLACEHOLDER PARAMETER DESCRIPTION}
}
\description{
Set Node Mapping
Function to set shape and border color according to node type
}
