% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeGraphingFunctions.R
\name{strip.cy.goo}
\alias{strip.cy.goo}
\title{Strip Cytoscape Graphing Goo
Function to extract node names from, e.g.: "ValidatedObjectAndEditString: validatedObject=ERBB3, editString=null"}
\usage{
strip.cy.goo(test)
}
\arguments{
\item{test}{PLACEHOLDER PARAMETER DESCRIPTION}
}
\description{
Strip Cytoscape Graphing Goo
Function to extract node names from, e.g.: "ValidatedObjectAndEditString: validatedObject=ERBB3, editString=null"
}
