% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCorrelationNetwork.R
\name{Visualize}
\alias{Visualize}
\title{Plotting helper function}
\usage{
Visualize(matrix)
}
\arguments{
\item{cccn_matrix}{A cccn_matrix like the one returned in MakeCorrelationNetwork}
}
\description{
This function takes a cccn_matrix as input and plots it
}
\examples{
Visualize(cccn_matrix)
}
